changes by comments
